requirements.txt
alembic==1.4.3
amqp==5.0.5
billiard==3.6.3.0
cachetools==4.2.4
celery==5.0.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
docxtpl==0.10.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.5.1
gittyleaks==0.0.31
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-spreadsheet==0.0.6
googleapis-common-protos==1.54.0
gunicorn==20.0.4
httplib2==0.20.2
idna==2.10
importlib-metadata==3.7.3
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
lxml==4.8.0
mailchimp-transactional==1.0.23
Mako==1.1.3
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.1
phonenumbers==8.12.8
prompt-toolkit==3.0.17
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.1
python-decouple==3.3
python-docx==0.8.11
python-editor==1.0.4
python-http-client==3.3.2
pytz==2021.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.8
scandir==1.10.0
sendgrid==6.6.0
sh==1.14.2
six==1.15.0
SQLAlchemy==1.4.39
starkbank-ecdsa==1.1.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.25.10
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
pandas==1.4.3
airtable==0.4.8
mailchimp-marketing==3.0.75
xlwt==1.3.0
pyramid==2.0
pathvalidate==2.5.0
scikit-learn==1.1.1
wwo-hist==0.0.7
stripe==4.2.0
# using in data-migration to read env variables
python-dotenv==0.20.0
PyJWT==2.6.0
Flask-CKEditor==0.4.6
